requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
drf-jwt==1.19.2
gunicorn==20.1.0
mysqlclient==2.1.1
phonenumbers==8.12.50
phonenumberslite==8.12.50
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
